feat: modularize kubernetes_cleanup_namespaces code
